chore(add tests): added tests
