+++ fixtures/tests/invalid_parser_configuration.yaml
@@ -1,0 +1,2 @@
+rootElements:
+  - Niccolo\DocparserPhp\Model\Parser\HTML\Element\InexistentParser
