+++ fixtures/tests/parser_configuration_with_duplicates.yaml
@@ -1,0 +1,3 @@
+rootElements:
+  - Niccolo\DocparserPhp\Model\Parser\HTML\Element\DoctypeParser
+  - Niccolo\DocparserPhp\Model\Parser\HTML\Element\DoctypeParser
